Fixed: comment_count is returning as Integer.

+++ models/articles.model.js
@@ -15,7 +15,7 @@ exports.fetchAllArticles = async (
     ) {
       return Promise.reject({ status: 404, message: "Topic not Found" });
     }
-    let query = `SELECT articles.article_id, articles.title, articles.author, articles.topic, articles.created_at, articles.votes, COUNT(comments.article_id) AS comment_count FROM articles
+    let query = `SELECT articles.article_id, articles.title, articles.author, articles.topic, articles.created_at, articles.votes, COUNT(comments.article_id)::Int AS comment_count FROM articles
     LEFT JOIN comments ON articles.article_id = comments.article_id`;
     if (topic === undefined) {
       query += ` GROUP BY articles.article_id ORDER BY ${sort_by} ${order};`;
